Bump Syncbox dependency

+++ src/core.rs
@@ -992,6 +992,6 @@ mod test {
 
     #[test]
     pub fn test_struct_sizes() {
-        assert_eq!(mem::size_of::<State>(), mem::size_of::<usize>());
+        assert_eq!(mem::size_of::<State>(), mem::size_of::<u64>());
     }
 }
